pyproject.toml
[project]
name = "discord-bridge"
version = "0.1.0"
description = "Discord Gateway API bridge that forwards events to HTTP endpoints"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "websockets>=12.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "structlog>=23.2.0",
]

[project.scripts]
discord-bridge = "discord_bridge.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
]

[tool.ruff]
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
